HW/final_project/racing/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.2
python-dotenv>=1.0.0
requests>=2.30.0
werkzeug>=3.0.0 